rename "personalization.rs" command file to "color.rs"
Seeing how large one command file is, I decided to have one command (including sub-commands) per file
