+++ zola/src/commands/mod.rs
@@ -1,1 +1,1 @@
-pub mod personalization;
+pub mod color;
